requirements.txt
ipython==9.0.2
langchain==0.3.21
langchain-chroma==0.2.2
langchain-community==0.3.20
langchain-core==0.3.49
langchain-google-genai==2.1.2
langgraph==0.3.21
PyYAML==6.0.2
streamlit==1.44.0